pyproject.toml
[project]
name = "memoryllm"
version = "0.1.0"
description = "A proxy to OpenAI compatible LLMs that injects user's memories"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.1",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
]
